Add SlottedStructure Component
It's a MESS, what with the various hidden component dependencies I've
encoded in this code. Clearly, I should merge the SlottedStructure and
Slotted classes, and also introduce 'dependency' semantics in to say "If
you're slottable you need to be able to have internal structure somehow"
but that relies on specific classes. eeeeurgh