Add component+event+query
I, uh, I'm not sure this is sane! If you want to find the position of the
entity you have to go and construct a query object, then tell the entity
to resolve it, then cast the output back.

That...seems way worse than just saying "Can you give me the Position
component". Hmm. HMM.