Implement tiebreakers in Schedule_RoundRobin
Tiebreakers are just more matches that keep being played until the tie
becomes unstuck. Uh, sorta inelegant, and the code is horrifying! - but it
will serve for now.

You do have the theoretical problem of an infinite 3-way tiebreaker
though. I might want to add a "Cap at like 20 rounds because this is
silly!" escape valve.