Defer Mountable "Destroyed?" to Mount
This is because Mountables don't actually take damage. I'm possibly going
to model this later, because you could theoretically shoot something while
it's holstered to damage it so they couldn't use it - but the holster
itself would take damage, and if destroyed, also destroy the weapon.
So...not right now necessary.