Add IDisplay
In the proposed system, we will have a "Current IDisplay" value. This
value will go through the following steps:

Render & render sub-consoles
Poll input & operate on input
Poll currentDisplay.NextDisplay() and set the next display to that value

This will mean that each IDisplay could be considered a state machine node
and if it has internal state, it should keep it on its own - so "IDisplay"
is really on reflection not the best name.

I have very little memory of Best Practices in the UI area, which is a
pity. Downside of server development I guess. On the other hand I don't
have to deal with UIs, which is a big upside!