Optional "Require"/"Must Be After" in Component
I can't really think of a great way to handle the Ordering Problem yet -
that is, the order of evaluation of Components is incredibly important and
you want a good way to say "Evaluate this one first!". I settled with
allowing Components to say "Uh oh I should be before/after this other one
something's wrong" but that will be incredibly awkward around the point at
which I start adding/removing components dynamically.