Run IsMet off GameQuery_Command
GameQuery_Command carries all the state you need to operate off of to
decide what to do next. Instead of storing the state in the Guidebook,
instead we can use the query to pass it in.

This is kind of confusing because GameQuery_Command doesn't express in
its naming "I am asking for a Command based on the info in this object"
very well. Let me think on the naming.

Also, I'm working out of a library now and these high schoolers won't
stop talking, it's really quite vexing.