Add distance option to Condition_Distance
This is to support "If you are further away than you weapon's max range,
DO X" or "If you are in melee range, DO Y" or similar constructs.

However, it's a little problematic in that if I want to randomly pick a
condition, now I have to cross-product my n different conditions within
Condition_Distance. Furthermore, it has issues with "Who's the target?"

I'm considering making the Target a piece of state that carries through
the condition evaluations and is switched by an explicit SingleClause,
sort of like how OpenGL does the drawing calls, but that's a little
further off.