Rename Action->AIAction, GeneFactory->GeneListing
This is because I was reading the Genetic code and it uses the Action
class (the one which indicates a procedure) and I realized that this was
super confusing!

Also the GeneFactory didn't actually produce anything, it just listed
them. I guess there should be a note there - Genes should be
Immutable. Unfortunately there's no IImmutable, or Immutable base class
in C#, which is unfortunate.