Add slow knife, fast pistol mechs
If I give the learning mech the fast pistol, it should end up with an AI
that always runs away and fires, thereby giving a perfect win.

This appears to be what happens! Also, that script is very tiny compared
to the amount of script space I gave to the AI.