Collapse Component_AttriuteModifier to one class
There were two versions - Mult and Flat. Collapsed into one. Next, I
will add conditions (only + to items of a certain label, for example)
which will be the basis of "Pilot has +3 to To Hit for Rifles".

To Hit. Heh. I should do THACO!

No, that's a terrible idea and I should feel bad for thinking it.