Only paint structure if base structure present
This is because it was painting Weapon structures, but it would use the
container structure as the weapon structure value. Currently, weapons
are actually invincible and have NO structure, because of that fact, so
it caused a error when drawing.

Fixed error by avoiding painting structure when the component does not
register a base structure.