Add ability to assign pilot to mech
I am considering making it a GameEvent, but in the current
conceptualization of the game the pilot will be more or less tightly
coupled to the mech - they'll be assigned once only, and then left there
forever. So swappability in a nice way is not anticipated to be a need.