Hook evolve command into main menu
It's under E

Also, it doesn't...work. It never hits a solution that beats the enemy.
So, it runs, but doesn't optimize.

I need to figure out what's goin' on here, but I suspect that the
equipment clauses are taking up so much of it that it squeezes out the
actions. Hmm.

Currently takes about 20 minutes to run on my machine.