Fix evolver option from Main Menu
It did not previously build the mechs and attach weapons, which lead to
mechs always losing because they'd exit out on the first round of combat
due to combat-ineffectiveness.

Fixed by making the Step_AttachAttachable's enumeration actually return
the correct class and attach to the working item in the apply step
function.