Add CommandStub for PrepareAttack
Essentially, I need to have "atomic actions" that I can apply in-order, so
that when you go into focus you can queue them up and then execute them
all at once. The GameEvents themselves are explicitly coupled to the
timings and specific state of the arena, however - if you queue up an
attack, a weapon swap, and another attack, it would end up attacking twice
with the old weapon because it'd use the state at time of queue, not at
time of resolution.

So I'm doing something convoluted with stubs, where you queue the stubs,
then it fills in the specific state at resolution time. I hope this works.