Make focus readout human-readable
The stubs now implement ToString() though I am not happy with how the
label for the attack is implemented! I should have just kept the input
as entity and harvested the IDs from there, because now it'll be easy to
accidentally mis-order the strings.

oh welp 4 days left TimE preSSURE